code refactoring, UI revamp, and bug fixes
