+++ src/app/material/material.module.ts
@@ -15,6 +15,7 @@ import { MatNativeDateModule } from '@angular/material/core';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatChipsModule } from '@angular/material/chips';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 import { MtxTooltipModule } from '@ng-matero/extensions/tooltip';
@@ -36,6 +37,7 @@ let MATERIAL_MODULES = [
   MatTooltipModule,
   MatDialogModule,
   MatChipsModule,
+  MatCheckboxModule,
 ];
 
 @NgModule({
