--- src/app/modal/calendar/calendar.component.sass
@@ -1,35 +1,0 @@
-::ng-deep mat-datepicker-toggle mat-datepicker
-    width: 10px !important
-    height: 30px !important
-
-.container
-    height: auto
-
-    &__form
-        &__card
-            &__content
-                &__div
-                    display: flex
-                    justify-content: center
-
-                    &__header
-                        display: flex
-                        justify-content: center
-                        margin-top: 5%
-                        margin-bottom: 10px
-                        font-size: large
-
-                    &__form-field
-                        width: 100%
-
-            &__actions
-                display: flex
-                justify-content: center
-                &__button
-                    width: 4%
-                    margin: 5px 0 0 10px !important
-                    position: relative
-                    z-index: 0
-                    padding: 0.2em
-                    margin: -0.05em
-                    margin-left: 6px
