--- src/app/modal/issues/issues.component.html
@@ -1,1 +1,0 @@
-<p>issues works!</p>
