--- src/app/modal/repos/repos.component.html
@@ -1,1 +1,0 @@
-<p>repos works!</p>
