--- src/app/modal/resume/resume.component.html
@@ -1,3 +1,0 @@
-<div class="container">
-  <iframe class="container__iframe" src="assets/files/resume.pdf"></iframe>
-</div>
