+++ src/app/pages/home/about/about.component.cy.ts
@@ -1,0 +1,7 @@
+import { AboutComponent } from './about.component'
+
+describe('AboutComponent', () => {
+  it('should mount', () => {
+    cy.mount(AboutComponent)
+  })
+})
