+++ src/app/pages/home/about/about.component.sass
@@ -1,0 +1,52 @@
+.container
+    display: flex
+    justify-content: center
+    &__card
+        flex-direction: row
+        margin: .5rem .25rem 0 .2rem
+        border-radius: .5rem
+        width: 90vw
+
+        &:hover
+            box-shadow: .02rem .5rem .4rem .025rem #ffd740
+
+@media screen and (max-width: 768px)
+    .container
+        &__card
+            flex-direction: column
+
+::ng-deep mat-datepicker-toggle mat-datepicker
+    width: 10px !important
+    height: 30px !important
+
+.message_container
+    height: auto
+
+    &__form
+        &__card
+            &__content
+                &__div
+                    display: flex
+                    justify-content: center
+
+                    &__header
+                        display: flex
+                        justify-content: center
+                        margin-top: 5%
+                        margin-bottom: 10px
+                        font-size: large
+
+                    &__form-field
+                        width: 100%
+
+            &__actions
+                display: flex
+                justify-content: center
+                &__button
+                    width: 4%
+                    margin: 5px 0 0 10px !important
+                    position: relative
+                    z-index: 0
+                    padding: 0.2em
+                    margin: -0.05em
+                    margin-left: 6px
