+++ src/app/pages/home/activity/activity.component.cy.ts
@@ -1,0 +1,7 @@
+import { ActivityComponent } from './activity.component'
+
+describe('ActivityComponent', () => {
+  it('should mount', () => {
+    cy.mount(ActivityComponent)
+  })
+})
