+++ src/app/pages/home/projects/projects.component.cy.ts
@@ -1,0 +1,7 @@
+import { ProjectsComponent } from './projects.component'
+
+describe('ProjectsComponent', () => {
+  it('should mount', () => {
+    cy.mount(ProjectsComponent)
+  })
+})
