+++ src/app/pages/home/projects/projects.component.html
@@ -1,0 +1,98 @@
+<div class="container">
+  <div style="width: 90vw">
+    <mat-card class="container__card container__other-cards">
+      <mat-card-header class="container__card-header">Projects</mat-card-header>
+      <div
+        style="
+          height: 90%;
+          width: 90%;
+          justify-content: center;
+          display: flex;
+          align-items: center;
+        "
+      >
+        <mat-card-content style="margin: 0.2rem 0 0 0; padding: 0">
+          <table>
+            <tr>
+              <td>Work related:</td>
+              <td>
+                <ul>
+                  <li>
+                    Consuming REST APIs from SCM/ALM/Code Quality tools in Go
+                    plugins
+                  </li>
+                  <li>
+                    Parsing and pushing data to internal microservices, using
+                    GraphQL REST endpoints
+                  </li>
+                  <li>
+                    Using more GraphQL in Angular web application to fetch data
+                    from internal microservices
+                  </li>
+                  <li>
+                    Using charting libraries, internal NPM packages and Angular
+                    Material for data visualisation in Angular web application
+                  </li>
+                  <li>
+                    Offloading core web application by creating Angular Micro
+                    Frontends
+                  </li>
+                  <li>
+                    Ensuring code quality with semantic analysis, security
+                    analysis, SAST & DAST, using tools like SonarQube, Fortify
+                    SAST, Synopsys BlackDuck, GHAS, CodeQL, etc.
+                  </li>
+                </ul>
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <a [href]="portfolioRepo">Portfolio</a>
+              </td>
+              <td>
+                <ul>
+                  <li>Using Go and MongoDB to implement server</li>
+                  <li>
+                    Using Go and MongoDB to implement integrations from GitHub
+                  </li>
+                  <li>
+                    Using Angular, Angular Material and ngx-charts to develop UI
+                  </li>
+                  <li>
+                    Using Docker and Docker Compose, and DigitalOcean droplets
+                    to containerise MongoDB and backend services, and nginx
+                    reverse proxy to deploy UI on Linux based container
+                  </li>
+                  <li>
+                    Using CircleCI and GitHub Actions with tools like
+                    golangci-lint, Cypress, ESLint and SonarQube Cloud for code
+                    quality checks
+                  </li>
+                </ul>
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <a [href]="githubOrg"> Ecommerce website clone </a>
+              </td>
+              <td>
+                <ul>
+                  <li>Using Angular to clone Amazon landing page</li>
+                  <li>
+                    Developing backend services and microservices will involve
+                    NodeJS, Go and Java SpringBoot and Hibernate.
+                  </li>
+                  <li>
+                    Co-developer:
+                    <a [href]="linkedIn" target="_blank">@Aman Verma</a>
+                  </li>
+                  <li>Stay tuned for updates</li>
+                </ul>
+              </td>
+            </tr>
+          </table>
+        </mat-card-content>
+      </div>
+    </mat-card>
+  </div>
+</div>
