+++ src/app/pages/pages.module.ts
@@ -4,17 +4,29 @@ import { HomeComponent } from './home/home.component';
 import { TodosComponent } from './todos/todos.component';
 import { ArticlesComponent } from './articles/articles.component';
 import { MaterialModule } from '../material/material.module';
-import { ModalModule } from '../modal/modal.module';
 import { SharedModule } from '../shared/shared.module';
 import { PagesRoutingModule } from './pages-routing.module';
 import { ChartComponent } from './chart/chart.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { AboutComponent } from './home/about/about.component';
+import { ProjectsComponent } from './home/projects/projects.component';
+import { ActivityComponent } from './home/activity/activity.component';
 
 @NgModule({
-  declarations: [HomeComponent, TodosComponent, ArticlesComponent, ChartComponent],
+  declarations: [
+    HomeComponent,
+    TodosComponent,
+    ArticlesComponent,
+    ChartComponent,
+    AboutComponent,
+    ProjectsComponent,
+    ActivityComponent,
+  ],
   imports: [
     CommonModule,
     MaterialModule,
-    ModalModule,
+    FormsModule,
+    ReactiveFormsModule,
     SharedModule,
     PagesRoutingModule,
   ],
