+++ src/app/skeleton/header/header.component.html
@@ -1,6 +1,5 @@
-<mat-toolbar color="accent" class="container">
+<mat-toolbar class="container">
   <mat-toolbar-row class="container__name">{{ name }}</mat-toolbar-row>
-
   <mat-toolbar-row class="container__details">{{
     headerDetails
   }}</mat-toolbar-row>
