+++ src/app/skeleton/header/header.component.sass
@@ -1,12 +1,18 @@
 .container
     min-width: 100vw
+    background-image: url(~src/assets/images/header.webp)
+    background-size: 150%
+    background-position: center
+    background-repeat: no-repeat
     &__name
         font-size: 2rem
         justify-content: center
+        color: #e1d3d3bd
     &__details
         justify-content: center
-        font-size: 0.95rem
-        margin-top: -1rem
+        font-size: 1.2rem
+        margin-top: -.8rem
+        color: #e1d3d3bd
 
     @media screen and (max-width: 535px)
         .container
