+++ src/app/skeleton/menu/menu.component.html
@@ -1,6 +1,6 @@
-<mat-toolbar class="container" color="accent">
+<mat-toolbar class="container stars">
   <a href="https://blog.ssnk.in" target="_blank" class="container__button"
-    >articles</a
+    >blog</a
   >
   <button
     class="container__button"
@@ -15,6 +15,6 @@
     routerLinkActive="active"
     ariaCurrentWhenActive="page"
     class="container__button"
-    >todos</a
+    >admin</a
   >
 </mat-toolbar>
