+++ src/app/skeleton/menu/menu.component.sass
@@ -4,13 +4,33 @@
     display: flex
     min-width: 100vw
 
-    &:hover
-        background-color: #f5f5f5
-
     &__button
         text-decoration: none
         cursor: pointer
-        color: rgba(0,0,0,.87)
+        mix-blend-mode: screen
+        color: #9e9e9e
 
         &:hover
-            color: #e65c11
+            color: #ffd740
+
+.stars
+    position: relative
+    background-color: #000
+    overflow: hidden
+
+    &::before
+        content: ""
+        position: absolute
+        top: -50%
+        left: -50%
+        width: 200%
+        height: 200%
+        background-image: radial-gradient(white 0%, transparent 80%)
+        animation: stars 10s ease-in-out infinite
+
+@keyframes stars
+    from
+        transform: translate(0, 0)
+
+    to
+        transform: translate(-50%, -50%)
